user_service/requirements.txt
azure-functions
jwt